mejoras

+++ droid-appquejas-ws/src/main/java/com/pe/droid/appquejas/domain/Cliente.java
@@ -16,7 +16,7 @@ public class Cliente implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="id_cliente")
-	private int idCliente;
+	private Long idCliente;
 
 	@Column(name="numero_departamento")
 	private String numeroDepartamento;
@@ -34,11 +34,11 @@ public class Cliente implements Serializable {
 	public Cliente() {
 	}
 
-	public int getIdCliente() {
+	public Long getIdCliente() {
 		return this.idCliente;
 	}
 
-	public void setIdCliente(int idCliente) {
+	public void setIdCliente(Long idCliente) {
 		this.idCliente = idCliente;
 	}
